% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chemCheck.R
\name{chemCheck}
\alias{chemCheck}
\title{Function to flag samples if basic chemistry is unreasonable}
\usage{
chemCheck(qw.data)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag samples if basic chemistry is unreasonable
}

