% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/converTime.R
\name{convertTime}
\alias{convertTime}
\title{Companion function to readNWISodbc to convert dates to proper time zone}
\usage{
convertTime(datetime, timezone, daylight)
}
\description{
Companion function to readNWISodbc to convert dates to proper time zone
}

