% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flagReport.R
\name{flagReport}
\alias{flagReport}
\title{flagReport}
\usage{
flagReport(DSN, env.db = "01", qa.db = "02", STAIDS,
  begin.date = Sys.Date() - 365 * 10, end.date = Sys.Date(),
  outputFile = "flagReport.pdf", outputDir = "output")
}
\arguments{
\item{DSN}{A character string containing the DSN for your local server}

\item{env.db}{A character string containing the database number of environmental samples}

\item{qa.db}{A character string containing the database number of QA samples}

\item{STAIDS}{A character vector of stations IDs}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd). Default is 10 years from current date.}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd). Default is current date.}

\item{outputFile}{Filename of report output.}

\item{outputDir}{Directory to store output files.}
}
\description{
Pulls data from internal NWIS server and runs automated data checks, generates tables, and produces
an indexed notebook of flagged samples, tables, and plots for all sites entered.
}

