% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/historicCheck.R
\name{historicCheck}
\alias{historicCheck}
\title{Function to flag samples if values are out of range of historic data}
\usage{
historicCheck(qw.data)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag samples if values are out of range of historic data
}

