% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needsReview.R
\name{needsReview}
\alias{needsReview}
\title{Function to return sample needing review}
\usage{
needsReview(qw.data, flaggedResults = NULL)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{flaggedResults}{Optional dataframe of flagged results generated from \code{flagSummary} function}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to return sample needing review
}
\details{
Returns list of unapproved results (DQI of I, S, or P) and if the result has been flagged anywhere
}
\examples{
data("exampleData",package="WQReview")
pestCheckOut <- pestCheck(qw.data=qw.data,
             returnAll = FALSE)
}
