% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pestCheck.R
\name{pestCheck}
\alias{pestCheck}
\title{Function to flag pesticide samples}
\usage{
pestCheck(qw.data, returnAll = FALSE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag pesticide samples
}
\details{
Compares each sample with DQI code of "I","S", or "P" to ranges of all prior approved data ("R","O","A"),
and flags samples that are suspisciously high or low i nnumber of hits for pesticide schedule 2437.
Definitions of checks can be found at http://internal.cida.usgs.gov/NAWQA/data_checks/docs/files/check30-sql.html
}
\examples{
data("exampleData",package="WQReview")
pestCheck(qw.data=qw.data,
             returnAll = FALSE)
}

