% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/publicData.R
\name{publicData}
\alias{publicData}
\title{Function to pull data from NWIS}
\usage{
publicData(STAIDS, dl.parms = "All", parm.group.check = TRUE,
  begin.date = "", end.date = "")
}
\arguments{
\item{STAIDS}{A character vector of stations IDs to pull data}

\item{dl.parms}{A character vector of pcodes to pull data}

\item{parm.group.check}{A logical of weather or not to use NWIS parameter groups. If TRUE, must use NWIS parameter group names in dl.parms}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd)}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd)}
}
\description{
Pulls data from NWIS WEB servers using functions from the dataRetrieval package.
}

