% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualitySummary.R
\name{qualitySummary}
\alias{qualitySummary}
\title{qualitySummary}
\usage{
qualitySummary(qw.data, reports = NULL)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}
}
\description{
Takes output data object from readNWISodbc and generates a summary of data quality.
}

