% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{qw.data}
\alias{qw.data}
\title{Output qw.data object for the Big Thompson River generated using readNWISodbc.}
\source{
Colorado NWIS server
}
\usage{
qw.data
}
\description{
This contains all data for the Big Thompson River site number 06733000 pulled from NWIS using the readNWISodbc function. 
This dataset is used for all run examples.
}
\author{
Joe Mills, 2015-11-02
}
