% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwblankPlot.R
\name{qwblankPlot}
\alias{qwblankPlot}
\title{Blank sample timeseries}
\usage{
qwblankPlot(qw.data, plotparm, site.selection = "All", printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{plotparm}{A character vector of the parameter to plot.}

\item{site.selection}{A character vector of site IDs to plot}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a boxplot of blank results
}
\examples{
data("exampleData",package="WQReview")
qwblankPlot(qw.data = qw.data,
                       site.selection = "All",
                       plotparm = "00915",
                       printPlot = TRUE
                       )
}
