% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qwmapPlot.R
\name{qwmapPlot}
\alias{qwmapPlot}
\title{Parameter matrix plot}
\usage{
qwmapPlot(qw.data, site.selection, plotparm, map = NULL, begin.date = NULL,
  end.date = NULL)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of parameters to plot}

\item{map}{A ggmap object to plot data on, not required}

\item{begin.date}{Data range to subset data to}

\item{end.date}{Data range to subset data to}
}
\description{
Takes output data object from readNWISodbc and prints a parameter matrix plot
}

