% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qwmatrixPlot.R
\name{qwmatrixPlot}
\alias{qwmatrixPlot}
\title{Parameter matrix plot}
\usage{
qwmatrixPlot(qw.data, site.selection, plotparm)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of parameters to plot}
}
\description{
Takes output data object from readNWISodbc and prints a parameter matrix plot
}

