% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwOutlier.R
\name{qwOutlier}
\alias{qwOutlier}
\title{Multivariate outlier detection
Takes output data object from readNWISodbc and runs multivariate outlier tests
using the robust mahalanobis distance.}
\usage{
qwOutlier(qw.data, quantile)
}
\arguments{
\item{qw.data}{A qw.data data object produced from readNWISodbc}

\item{quantile}{The quantile to use for outlier cutoff}
}
\description{
Multivariate outlier detection
Takes output data object from readNWISodbc and runs multivariate outlier tests
using the robust mahalanobis distance.
}
\details{
Samples ar evaluated for the presence of multivariate outliers using the robust mahalnobis distance.
Parameters are grouped by expected relationships. For example, discharge major cations, SO4, and alkalinity are expected
to be coorlleated with each other due to similar geochemical behavoir of ions and relationships with conductance and discharge.
Similarly, nutrient samples are grouped because of common sources. Pesticide and trace metal data are not evaluated.
The robust mahalanobis distance is calculated using the covMcd function from the robustbase package.
Samples that have mahalnobis distances greater than the specified quantile are flagged as possible outliers.
If the sample has been reviewed and approved based on DQI code, the outlier flag is removed.
WORK IN PROGRESS
}

