% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qwparmBox.R
\name{qwparmBoxPlot}
\alias{qwparmBoxPlot}
\title{Parameter boxplot}
\usage{
qwparmBoxPlot(qw.data, site.selection, plotparm, facet = "multisite",
  new.threshold = 60 * 60 * 24 * 30, show.points = FALSE,
  log.scale = FALSE, highlightrecords = NULL, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of pcodes to plot}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.points}{Show samples points}

\item{log.scale}{Plot y axis on a log scale}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}
}
\description{
Takes output list from readNWISodbc and prints a boxplot of parameters
}

