% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwparmparmPlot.R
\name{qwparmParmPlot}
\alias{qwparmParmPlot}
\title{Parameter-parameter plot}
\usage{
qwparmParmPlot(qw.data, site.selection, xparm, yparm, facet = "multisite",
  new.threshold = 60 * 60 * 24 * 30, show.lm = FALSE, log.scaleY = FALSE,
  log.scaleX = FALSE, highlightrecords = " ", wySymbol = FALSE,
  labelDQI = FALSE, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{xparm}{Character string of parameter to plot on x axis}

\item{yparm}{Character string of parameter to plot on y axis}

\item{facet}{Character string of either "multisite" for plotting all sites on one plot or "Facet" for plotting sites on individual plots}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.lm}{Add a linear fit to plot}

\item{log.scaleY}{Logical. Plot y parameter on a log scale.}

\item{log.scaleX}{Logical. Plot x parameter on a log scale.}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{wySymbol}{Make current water-year highlighted.}

\item{labelDQI}{Logical. Should points be labeled with DQI code.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and returns a plot of parameter vs parameter.
}
\examples{
data("exampleData",package="WQReview")
qwparmParmPlot(qw.data = qw.data,
              site.selection = "06733000",
              xparm = "00915",
              yparm = "00061",
              facet = "multisite",
              new.threshold = 60*60*24*30,
              show.lm=FALSE,
              log.scaleY = FALSE,
              log.scaleX = FALSE,
              highlightrecords = NULL,
              wySymbol = FALSE,
              labelDQI = FALSE,
              printPlot = TRUE)
}

