% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwrepBoxPlot.R
\name{qwrepBoxPlot}
\alias{qwrepBoxPlot}
\title{Boxplot of replicate relative percent difference}
\usage{
qwrepBoxPlot(qw.data, site.selection, plotparm, new.threshold = 60 * 60 * 24 *
  30, show.points = FALSE, highlightrecords = NULL, new.reps = Sys.time()
  - 60 * 60 * 24 * 30, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of pcodes to plot}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.points}{Show samples points}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{new.reps}{The threshold value in seconds from current system time for "new" replicate samples.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a boxplot of replicate RPD
}
\examples{
data("exampleData",package="WQReview")
qwrepBoxPlot(qw.data = qw.data,
                       site.selection = "06733000",
                       plotparm = "00915",
                       new.threshold = 60*60*24*30,
                       show.points = FALSE,
                       highlightrecords = NULL,
                       new.reps = Sys.time() - 60*60*24*30,
                       printPlot = TRUE)
}
