% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwrepPlot.R
\name{qwrepPlot}
\alias{qwrepPlot}
\title{Boxplot of replicate relative percent difference}
\usage{
qwrepPlot(qw.data, plotparm, site.selection = "All", printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{plotparm}{A character vector of pcodes to plot}

\item{site.selection}{A character vector of site IDs to plot}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a boxplot of replicate RPD
}
\examples{
data("exampleData",package="WQReview")
qwrepPlot(qw.data = qw.data,
                       site.selection = "06733000",
                       plotparm = "00915",
                       printPlot = TRUE)
}
