% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwscSumPlot.R
\name{qwscSumPlot}
\alias{qwscSumPlot}
\title{Sum ions vs conductance plot}
\usage{
qwscSumPlot(qw.data, site.selection, facet = "multisite", new.threshold = 60
  * 60 * 24 * 30, highlightrecords = NULL, wySymbol = FALSE,
  printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{facet}{Character string of either "multisite" for plotting all sites on one plot or "Facet" for plotting sites on individual plots}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{wySymbol}{Make current water-year highlighted.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a plot of sum ions vs. conductance. Requires charge balance = TRUE in NWISPullR
}
\examples{
data("exampleData",package="WQReview")
qwscSumPlot(qw.data = qw.data,
              site.selection = "06733000",
              facet = "multisite",
              new.threshold = 60*60*24*30,
              highlightrecords = NULL,
              wySymbol = FALSE,
              printPlot = TRUE)
}

