% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwtsPlot.R
\name{qwtsPlot}
\alias{qwtsPlot}
\title{Parameter timeseries plot}
\usage{
qwtsPlot(qw.data, site.selection, plotparm, new.threshold = 60 * 60 * 24 * 30,
  show.q = FALSE, show.smooth = FALSE, highlightrecords = " ",
  facet = "multisite", wySymbol = FALSE, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of parameters to plot}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.q}{Logical to plot instantaneous hydrograph}

\item{show.smooth}{Logical to add a loess smooth to plot}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{facet}{Logical. Should sites be plotted together on one plot or on separate plots using faceting.}

\item{wySymbol}{Make current water-year highlighted.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a plot of parameter vs time.
}
\examples{
data("exampleData",package="WQReview")
qwtsPlot <- function(qw.data = qw.data,
                       site.selection = "06733000",
                       plotparm = "00915",
                       new.threshold = 60*60*24*30,
                       show.q = FALSE,
                       show.smooth = FALSE,
                       highlightrecords = " ",
                       facet = "multisite",
                       wySymbol = FALSE,
                       printPlot = TRUE)
}

