<<<<<<< HEAD
% Generated by roxygen2 (4.1.0): do not edit by hand
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
>>>>>>> df8b1f1353b0ab7d8e77b196e915c7dd549625ff
% Please edit documentation in R/readNWISodbc.R
\name{readNWISodbc}
\alias{readNWISodbc}
\title{Function to pull data from NWIS}
\usage{
readNWISodbc(DSN = NULL, env.db = "01", qa.db = "02", STAIDS = NULL,
  dl.parms = "All", parm.group.check = TRUE, begin.date = NA,
  end.date = NA)
}
\arguments{
\item{DSN}{A character string containing the DSN for your local server}

\item{env.db}{A character string containing the database number of environmental samples}

\item{qa.db}{A character string containing the database number of QA samples}

\item{STAIDS}{A character vector of stations IDs}

\item{dl.parms}{A character vector of pcodes to pull data, e.g. "00300"}

\item{parm.group.check}{A logical of weather or not to use NWIS parameter groups. If TRUE, must use NWIS parameter group names in dl.parms}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd)}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd)}
}
\description{
Pulls data from NWIS internal servers using ODBC connection and returns a list containing dataframes.
}

