% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{reports}
\alias{reports}
\title{Output reports object for the Big Thompson River generated using various report functions.}
\source{
Colorado NWIS server
}
\usage{
qw.data
}
\description{
This contains all data for the Big Thompson River site number 06733000 pulled from NWIS using the readNWISodbc function.
This dataset is used for all run examples.
}
\author{
Joe Mills, 2015-11-02
}

