% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repTabler.R
\name{repTabler}
\alias{repTabler}
\title{repTabler}
\usage{
repTabler(qw.data)
}
\arguments{
\item{qw.data}{A qw.data object generater by NWIS_PullR}
}
\description{
Generates a table of environmental samples and replicates and calculates relative percent difference
}

