% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srsSummary.R
\name{srsSummary}
\alias{srsSummary}
\title{Summarize SRS sample agreement}
\usage{
srsSummary(srsValues, srsResults, reportingLevels)
}
\arguments{
\item{srsValues}{A dataframe containing the SRS values. The first column contains the SRS sample ID,the second column contains the SRS analyte parameter code, the third column contains the SRS analyte name, and the fourth column contains the SRS MPV}

\item{srsResults}{A dataframe containing the results of the SRS analyses from the lab. The first column containssample descriptions (e.g. date and time), the second column contains the SRS sample ID, and the remaining columns containthe SRS results. SRS Result columns must be named with the cooresponding paramter code.}

\item{reportingLevels}{A dataframe containing the reporting levels for each analyte. The first column contains the parameter code and the second column contains hte associated reporting level}
}
\description{
Summarizes standard reference sample (SRS) results and automatically flags bad SRS results.
}

