% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/templateExport.R
\name{templateExport}
\alias{templateExport}
\title{templateExport exports templates for use with the nwisupload function}
\usage{
templateExport()
}
\arguments{
\item{There}{are no arguments to this function. Just run to export templates}
}
\description{
exports templates for use with the nwisupload function. Files exported to working directory as tab-delimited files.
}

