% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterYear.R
\name{waterYear}
\alias{waterYear}
\title{Stolen from smwrBase to remove dependency on that package}
\usage{
waterYear(x, numeric = FALSE)
}
\description{
Stolen from smwrBase to remove dependency on that package
}

