% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whatData.R
\name{whatData}
\alias{whatData}
\title{Look for missing data for a set of samples}
\usage{
whatData(qw.data, searchParms, begin.date = NULL, end.date = NULL,
  groups = FALSE)
}
\arguments{
\item{qw.data}{A qw.data object generater by readNWISodbc}

\item{searchParms}{A character vector of parameter codes to search for if groups = FALSE, otherwise a dataframe of parameter codes in character format to search for grouped by named columns if groups= TRUE. Maximum number of columns(groups) is 10}

\item{begin.date}{Subset data to begin date}

\item{end.date}{Subset data to end date.}

\item{groups}{Logical. Search for searchParms by groups instead of one long vector.}
}
\description{
Searches for missing data and returns a dataframe of missing data by record number.
}
\examples{
data("exampleData",package="WQReview")
searchParms <- data.frame(cations = c("00915","00930","00925","00935","01123"),
                          anions = c("00945","00940","99220",NA,NA),
                          nutrients = c("00608","00631",NA,NA,NA),
                          stringsAsFactors = FALSE)
whatData(qw.data = qw.data,
        searchParms = searchParms,
        begin.date = NULL,
        end.date = NULL,
        groups=TRUE)
}

