% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wholevPart.R
\name{wholevPart}
\alias{wholevPart}
\title{Part vs whole analysis}
\usage{
wholevPart(qw.data)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}
}
\description{
Takes output list from readNWISodbc and checks whole vs part concentrations
}
\examples{
data("exampleData",package="WQReview")
wholevPart(qw.data=qw.data)
}

