% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankSummary.R
\name{blankSummary}
\alias{blankSummary}
\title{blankSummary Generate a summary table of blank analyses}
\usage{
blankSummary(qw.data, STAIDS = "All", begin.date = NA, end.date = NA,
  multiple.levels = FALSE)
}
\arguments{
\item{qw.data}{a qw.data list generated from readNWISodbc}

\item{STAIDS}{A character vector of station IDs to summarize or All for all site ids in qw.data.Defaults to "All"}

\item{begin.date}{Character string (yyyy-mm-dd) of beggining date to subset blank analyses. Leave blank for all analyses.}

\item{end.date}{Character string (yyyy-mm-dd) of ending date to subset blank analyses. Leave blank for all analyses.}

\item{multiple.levels}{Logical to analyze multiple reporting levels separetly. Default is TRUE.}
}
\description{
Generates a summary table of blank analyses for all analytes in a qw.data list from readNWISodbc.
}
\examples{
data("exampleData",package="WQReview")
blankSummaryOut <- blankSummary(qw.data = qw.data,
             STAIDS = unique(qw.data$PlotTable$SITE_NO),
             begin.date = "2009-01-01",
             end.date = "2013-01-01",
             multiple.levels = FALSE)
}
