% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemCheck.R
\name{chemCheck}
\alias{chemCheck}
\title{chemCheck
Function to flag samples if basic chemistry is unreasonable}
\usage{
chemCheck(qw.data, returnAll = FALSE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
chemCheck
Function to flag samples if basic chemistry is unreasonable
}
\details{
Performs chemical checks for expected ranges of O2, pH, Sc, and chargebalance. Definitions of checks can be found at http://internal.cida.usgs.gov/NAWQA/data_checks/docs/files/check30-sql.html
}
\examples{
data("exampleData",package="WQReview")
chemCheckOut <- chemCheck(qw.data=qw.data,
             returnAll = FALSE)
}

