% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUpload.R
\name{dataUpload}
\alias{dataUpload}
\title{dataUpload formats files for upload into QWData}
\usage{
dataUpload(qwsampletype, labfile, pcodefile, DSN = "NWISCO", env.db = "01",
  qa.db = "02", writeFiles = FALSE, qwresultname = "",
  qwsamplename = "", qwsamplefile = "", qwsample.begin.date = NA,
  qwsample.end.date = NA, censor = FALSE, agencycode = "USGS",
  labid = NA, projectcode = NA, aquifercode = NA, analysisstatus = NA,
  analysissource = NA, hydrologiccond = "A", hydrologicevent = 9,
  tissue = NA, bodypart = NA, labcomment = NA, fieldcomment = NA,
  timedatum = NA, datumreliability = NA, colagencycode = "USGS")
}
\arguments{
\item{qwsampletype}{Numeric of type of qwsample file. Options 1: Pull qwsample form NWIS, 2:QWSample from file, 3: batch qwsample file}

\item{labfile}{Character string containing path to the labfile, must be .csv format}

\item{pcodefile}{Character string containing path to the pcodefile, must be .csv format}

\item{DSN}{Character string containing local NWIS server name}

\item{env.db}{Environmental database number}

\item{qa.db}{QA database number}

\item{writeFiles}{write out qwresult and qwsample files to text files in working directory}

\item{qwresultname}{Name of qwresult file}

\item{qwsamplename}{Name of qwsample file}

\item{qwsamplefile}{Character string containing path to the qwsample file if using qwsampletype = 2, must be .csv format}

\item{qwsample.begin.date}{Character string (yyyy-mm-dd) of beggining date to subset qwsample file to}

\item{qwsample.end.date}{Character string (yyyy-mm-dd) of ending date to subset qwsample file to}

\item{censor}{Logical. If TRUE results are censored to lrl provided in pcode file}

\item{agencycode}{Manual qwsample entry}

\item{labid}{Manual qwsample entry}

\item{projectcode}{Manual qwsample entry}

\item{aquifercode}{Manual qwsample entry}

\item{analysisstatus}{Manual qwsample entry}

\item{analysissource}{Manual qwsample entry}

\item{hydrologiccond}{Manual qwsample entry}

\item{hydrologicevent}{Manual qwsample entry}

\item{tissue}{Manual qwsample entry}

\item{bodypart}{Manual qwsample entry}

\item{labcomment}{Manual qwsample entry}

\item{fieldcomment}{Manual qwsample entry}

\item{timedatum}{Manual qwsample entry}

\item{datumreliability}{Manual qwsample entry}

\item{colagencycode}{Manual qwsample entry}
}
\description{
Assists in formatting data for upload into QWData. Use templateExport to generate spreadsheet templates to create labfile and pcode file.
}
