% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flagReport.R
\name{flagReport}
\alias{flagReport}
\title{flagReport}
\usage{
flagReport(stationFile, DSN, env.db = "01", qa.db = "02",
  begin.date = as.character(Sys.Date() - 365 * 10),
  end.date = as.character(Sys.Date()), outputDir)
}
\arguments{
\item{stationFile}{Filename of csv file containing network names in 1st column and station IDs in second column.}

\item{DSN}{A character string containing the DSN for your local server.}

\item{env.db}{A character string containing the database number of environmental samples.}

\item{qa.db}{A character string containing the database number of QA samples.}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd). Default is 10 years from current date.}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd). Default is current date.}

\item{outputDir}{Directory to store output files. Must be an absolute path, e.g. "D:/flagReports"}
}
\description{
Pulls data from internal NWIS server and runs automated data checks, generates tables, and produces
an indexed notebook of flagged samples, tables, and plots for all sites entered.
}

