% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flagSummary.R
\name{flagSummary}
\alias{flagSummary}
\title{Function to summarize data flags}
\usage{
flagSummary(qw.data)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}
}
\value{
A dataframe containing only flagged samples
}
\description{
Function to summarize data flags
}

