% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipDQI.R
\name{flipDQI}
\alias{flipDQI}
\title{flipDQI Flip DQI codes for given sites and parameters}
\usage{
flipDQI(STAIDS, records, parameters, dqiCodes, comments = NULL,
  commentType = NULL, DSN = "NWISCO", env.db = "01", qa.db = "02",
  keepRecordNo = FALSE, writeFiles = FALSE, qwresultname = "qwresult",
  qwsamplename = "qwsample")
}
\arguments{
\item{STAIDS}{A vector of station IDs}

\item{records}{A vector of record numbers for which DQI codes are to be flipped. Must be in the format as retrieved using readNWISodbc of record number with coorseponding database number appended with an underscore. E.g. "123456_01}

\item{parameters}{A vector of parameter codes of the same length as records for which DQI codes are to be flipped.See details}

\item{dqiCodes}{A vector of specified DQI codes of the same length as records.See details}

\item{comments}{An optional vector of comments for each result of the same length as records.See details}

\item{commentType}{An optional vector of comment types of the same length as records. Default is "F" for field. See details}

\item{DSN}{A character string containing the DSN for your local server}

\item{env.db}{Environmental database number}

\item{qa.db}{QA database number}

\item{keepRecordNo}{Logical. Retain record numbers in qwresult and qwsample. If TRUE, RECORD_NO column in qwresult and qwsample dataframes will need to be removed prior to batch upload to NWIS}

\item{writeFiles}{write out qwresult and qwsample files to text files in working directory}

\item{qwresultname}{Name of qwresult file}

\item{qwsamplename}{Name of qwsample file}
}
\description{
Generates batch files for input into QWData with user-specified DQI codes.
}
\details{
This function allows the user to generate batch files to change DQI codes for custom sets of parameters and record numbers.
The input vectors must be the same length, but record numbers can and often will be repeated. See example below for a demonstration of input.
}
\examples{
\dontrun{
#Will not run unless connected to NWISCO
STAIDS <- c("391517106223801","391500106224901")
records <- c("00306540_01","00306540_01","01305854_01","01305854_01")
parameters <- c("00400","00095","01106","01056")
dqiCodes <- c("R","R","Q","R")
flipDQI(STAIDS = STAIDS,
records = records,
parameter = parameters,
dqiCodes = dqiCodes,
DSN = "NWISCO",
env.db = "01",
qa.db = "02")
}

}
