% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicCheck.R
\name{historicCheck}
\alias{historicCheck}
\title{Function to flag samples if values are out of range of historic data}
\usage{
historicCheck(qw.data, returnAll = FALSE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Default is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag samples if values are out of range of historic data
}
\details{
Compares each sample with DQI code of "I","S", or "P" to ranges of all prior approved data ("R","O","A"),
and flags samples that are suspisciously high or low. 
Definitions of checks can be found at https://internal.cida.usgs.gov/NAWQA/data_checks/docs/files/check30-sql.html
}
\examples{
data("exampleData",package="WQReview")
historicCheckOut <- historicCheck(qw.data=qw.data,
             returnAll = FALSE)
}
