% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ionBalance.R
\name{ionBalance}
\alias{ionBalance}
\title{ionBalance}
\usage{
ionBalance(qw.data, wide = FALSE)
}
\arguments{
\item{qw.data}{A qw.data object generated from NWISPullR}
}
\description{
Takes output data object from NWISPullR and calculates a QWData standard ion charge balance.
}

