% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ionBalance.R
\name{ionBalance}
\alias{ionBalance}
\title{Calculate ion balance}
\usage{
ionBalance(qw.data, wide = FALSE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{wide}{Logical, return the ionbalance table in wide format.}
}
\description{
Takes output data object from readNWISodbc and calculates a standard ion charge balance.
}
\examples{
data("exampleData",package="WQReview")
ionBalanceOut <- ionBalance(qw.data=qw.data)
}
