% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pestCheck.R
\name{pestCheck}
\alias{pestCheck}
\title{Function to flag pesticide samples}
\usage{
pestCheck(qw.data)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag pesticide samples
}

