% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwblankPlot.R
\name{qwblankPlot}
\alias{qwblankPlot}
\title{Blank sample timeseries}
\usage{
qwblankPlot(qw.data, site.selection, plotparm, new.threshold = 60 * 60 * 24 *
  30, show.smooth = FALSE, highlightrecords = " ", facet = "multisite",
  scales = "fixed", wySymbol = FALSE, labelDQI = FALSE,
  printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of the parameter to plot.}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.smooth}{Logical to add a loess smooth to plot}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{facet}{Character string of either "multisite" for plotting all sites on one plot or "Facet" for plotting sites on individual plots}

\item{scales}{Character string to define y axis on faceted plots. Options are "free","fixed","free_x", or "free_y"}

\item{wySymbol}{Make current water-year highlighted.}

\item{labelDQI}{Logical. Should points be labeled with DQI code.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a timeseries plot of blanks
}
\examples{
data("exampleData",package="WQReview")
qwblankPlot(qw.data = qw.data,
                       site.selection = "06733000",
                       plotparm = "00915",
                       new.threshold = 60*60*24*30,
                       show.smooth = FALSE,
                       highlightrecords = " ",
                       facet = "multisite",
                       scales = "fixed",
                       wySymbol = FALSE,
                       labelDQI = FALSE,
                       printPlot = TRUE
                       )
}
