% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwblankPlot.R
\name{qwblankPlot}
\alias{qwblankPlot}
\title{Blank sample timeseries}
\usage{
qwblankPlot(qw.data, site.selection, plotparm, new.threshold = 60 * 60 * 24 *
  30, show.smooth = FALSE, highlightrecords = " ", facet = "multisite",
  wySymbol = FALSE, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data list generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of the parameter to plot.}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{wySymbol}{Make current water-year highlighted.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}

\item{begin.date}{Character string of begining date range to plot (yyyy-mm-dd)}

\item{end.date}{Character string of ending date range to plot (yyyy-mm-dd)}
}
\description{
Takes output data object from readNWISodbc and prints a timeseries plot of blanks
}
\examples{
data("exampleData",package="WQReview")
qwblankPlot <- function(qw.data = qw.data,
                       site.selection = "06733000",
                       plotparm = "00915",
                       new.threshold = 60*60*24*30,
                       show.q = FALSE,
                       show.smooth = FALSE,
                       highlightrecords = " ",
                       facet = "multisite",
                       wySymbol = FALSE,
                       printPlot = TRUE)
}

