% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwmatrixPlot.R
\name{qwmatrixPlot}
\alias{qwmatrixPlot}
\title{Parameter matrix plot}
\usage{
qwmatrixPlot(qw.data, site.selection, plotparm)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of parameters to plot}
}
\description{
Takes output data object from readNWISodbc and prints a parameter matrix plot
}
\examples{
data("exampleData",package="WQReview")
qwmatrixPlot(qw.data=qw.data, 
             site.selection = "06733000", 
             plotparm = c("00915","00945","00095")) 
}
