% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwparmBox.R
\name{qwparmBoxPlot}
\alias{qwparmBoxPlot}
\title{Parameter boxplot}
\usage{
qwparmBoxPlot(qw.data, site.selection, plotparm, facet = "multisite",
  new.threshold = 60 * 60 * 24 * 30, show.points = FALSE,
  log.scale = FALSE, highlightrecords = NULL, wySymbol = FALSE,
  labelDQI = FALSE, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of pcodes to plot}

\item{facet}{Character string of either "multisite" for plotting all sites on one plot or "Facet" for plotting sites on individual plots}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.points}{Show samples points}

\item{log.scale}{Plot y axis on a log scale}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{wySymbol}{Make current water-year highlighted.}

\item{labelDQI}{Logical. Should points be labeled with DQI code.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output list from readNWISodbc and prints a boxplot of parameters
}
\examples{
data("exampleData",package="WQReview")
qwparmBoxPlot(qw.data = qw.data,
              site.selection = "06733000",
              plotparm = c("00915","00095","00400"),
              facet = "multisite",
              new.threshold = 60*60*24*30,
              show.points = FALSE,
              log.scale = TRUE,
              highlightrecords = NULL,
              wySymbol = FALSE,
              labelDQI = FALSE,
              printPlot = TRUE)
}

