% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwparmparmPlot.R
\name{qwparmParmPlot}
\alias{qwparmParmPlot}
\title{Parameter-parameter plot}
\usage{
qwparmParmPlot(qw.data, site.selection, xparm, yparm, facet = "multisite",
  new.threshold = 60 * 60 * 24 * 30, show.lm = FALSE, log.scaleY = FALSE,
  log.scaleX = FALSE, highlightrecords = " ", print = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{xparm}{Character string of parameter to plot on x axis}

\item{yparm}{Character string of parameter to plot on y axis}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.lm}{Add a linear fit to plot}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}
}
\description{
Takes output data object from readNWISodbc and returns a plot of parameter vs parameter.
}

