% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qwrepBoxPlot.R
\name{qwrepBoxPlot}
\alias{qwrepBoxPlot}
\title{Boxplot of replicate relative percent difference}
\usage{
qwrepBoxPlot(reports, site.selection = "401723105400000",
  plotparm = "00915", new.threshold = 60 * 60 * 24 * 30,
  show.points = FALSE, highlightrecords = NULL, new.reps = Sys.time() - 60
  * 60 * 24 * 30)
}
\arguments{
\item{reports}{A reports object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of pcodes to plot}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{show.points}{Show samples points}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{log.scale}{Plot y axis on a log scale}
}
\description{
Takes output data object from readNWISodbc and prints a boxplot of replicate RPD
}

