% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwscSumPlot.R
\name{qwscSumPlot}
\alias{qwscSumPlot}
\title{Sum ions vs conductance plot}
\usage{
qwscSumPlot(qw.data, site.selection, facet = "multisite", new.threshold = 60
  * 60 * 24 * 30, highlightrecords = NULL, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{site.selection}{A character vector of site IDs to plot}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}

\item{plotparm}{A character string of ions to plot. Options are "Cations", "Anions", "Both"}
}
\description{
Takes output data object from readNWISodbc and prints a plot of sum ions vs. conductance. Requires charge balance = TRUE in NWISPullR
}

