% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qwseasonalPlot.R
\name{qwseasonalPlot}
\alias{qwseasonalPlot}
\title{Seasonal plot}
\usage{
qwseasonalPlot(qw.data, new.threshold = 60 * 60 * 24 * 30, site.selection,
  plotparm, facet = "multisite", show.q = FALSE, show.smooth = FALSE,
  highlightrecords = " ", print = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of parameters to plot}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{show.points}{A logical to show points on plot}

\item{log.scale}{A logical to plot y axis on log scale}
}
\description{
Takes output data object from readNWISodbc and prints a boxplot of a parameter by month.
}

