% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qwseasonalPlot.R
\name{qwseasonalPlot}
\alias{qwseasonalPlot}
\title{Seasonal plot}
\usage{
qwseasonalPlot(qw.data, new.threshold = 60 * 60 * 24 * 30, site.selection,
  plotparm, facet = "multisite", show.smooth = FALSE,
  highlightrecords = " ", wySymbol = FALSE, printPlot = TRUE)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{site.selection}{A character vector of site IDs to plot}

\item{plotparm}{A character vector of parameters to plot}

\item{facet}{Character string of either "multisite" for plotting all sites on one plot or "Facet" for plotting sites on individual plots}

\item{show.smooth}{Logical to add a loess smooth to plot}

\item{highlightrecords}{A character vector of record numbers to highlight in plot}

\item{wySymbol}{Make current water-year highlighted.}

\item{printPlot}{Logical. Prints plot to graphics device if TRUE}
}
\description{
Takes output data object from readNWISodbc and prints a boxplot of a parameter by month.
}
\examples{
data("exampleData",package="WQReview")
qwseasonalPlot(qw.data = qw.data,
               site.selection = "06733000",
               plotparm = "00095",
               facet = "multisite",
               new.threshold = 60*60*24*30,
               show.smooth = FALSE,
               highlightrecords = " ",
               wySymbol = FALSE,
               printPlot = TRUE)
}

