% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repTabler.R
\name{repTabler}
\alias{repTabler}
\title{Replicate agreement table}
\usage{
repTabler(qw.data)
}
\arguments{
\item{qw.data}{A qw.data object generated by readNWISodbc}
}
\description{
Generates a table of environmental samples and replicates and calculates relative percent difference
}
\examples{
data("exampleData",package="WQReview")
repTablerOut <- repTabler(qw.data=qw.data)
}

