% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{summaryStats}
\alias{summaryStats}
\title{Summary stats function}
\usage{
summaryStats(qw.data, STAIDS = NULL, parameterCd = NULL, begin.date,
  end.date)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{STAIDS}{A character vector of sites to summarize. Defaults to all sites in qw.data if unspecified.}

\item{parameterCd}{A character vector of parameter codes to summarize. Defaults to all pcodes if unspecified.}

\item{begin.date}{A character string of optional begin date range to summarize in "YYYY-MM-DD" format. Defaults to all if unspecified.}

\item{end.date}{A character string of optional end date range to summarize in "YYYY-MM-DD" format. Defaults to all if unspecified.}
}
\description{
Takes output data object from readNWISodbc and generates a table of summary statistcs
}

