% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{summaryStats}
\alias{summaryStats}
\title{Summary stats function}
\usage{
summaryStats(qw.data, STAIDS = NULL, parameterCd = NULL, begin.date = NA,
  end.date = NA)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{STAIDS}{A character vector of sites to summarize. Defaults to all sites in qw.data if unspecified.}

\item{parameterCd}{A character vector of parameter codes to summarize. Defaults to all pcodes if unspecified.}

\item{begin.date}{A character string of optional begin date range to summarize over in "YYYY-MM-DD" format. Defaults to all if unspecified.}

\item{end.date}{A character string of optional end date range to summarize over in "YYYY-MM-DD" format. Defaults to all if unspecified.}
}
\description{
Takes output data object from readNWISodbc and generates a table of summary statistcs
}
\examples{
data("exampleData",package="WQReview")
summaryStatsOut <- summaryStats(qw.data = qw.data,
             STAIDS="06733000",
             parameterCd = c("00095","00915","00945"),
             begin.date = "2010-01-01",
             end.date = "2011-12-31")
}

