% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaryStats.R
\name{summaryStats}
\alias{summaryStats}
\title{Sum ions vs conductance plot}
\usage{
summaryStats(qw.data, new.threshold = 60 * 60 * 24 * 30, parameterCd = NULL)
}
\arguments{
\item{qw.data}{A qw.data object generated from readNWISodbc}

\item{new.threshold}{The threshold value in seconds from current system time for "new" data.}

\item{parameterCd}{A character vector of parameter codes to summarize. Defaults to all pcodes if unspecified.}
}
\description{
Takes output data object from readNWISodbc and prints a plot of sum ions vs. conductance. Requires charge balance = TRUE in NWISPullR
}

