% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateExport.R
\name{templateExport}
\alias{templateExport}
\title{templateExport exports templates for use with the dataUpload function}
\usage{
templateExport()
}
\description{
exports templates for use with the nwisupload function. Files exported to working directory as tab-delimited files.
}
\examples{
templateExport()
}

