% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterYear.R
\name{waterYear}
\alias{waterYear}
\title{Stolen from smwrBase to remove dependency on that package
Takes a date and returns the water year}
\usage{
waterYear(x, numeric = FALSE)
}
\arguments{
\item{x}{an object of class "Date" or "POSIXt." Missing values are permitted and result in corresponding missing values in the output.}

\item{numeric}{numeric a logical value that indicates whether the returned values should be numeric TRUE or an ordered factor FALSE. The default value is FALSE.}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Stolen from smwrBase to remove dependency on that package
Takes a date and returns the water year
}
\examples{
waterYear("2015-01-01")
}

